+++ src/main/java/com/pix/domain/models/PixKey.java
@@ -17,9 +17,9 @@ public class PixKey {
     private String id;
     private String keyType;
     private String keyValue;
-    private BigDecimal agencyNumber;
+    private int agencyNumber;
     private String accountType;
-    private BigDecimal accountNumber;
+    private int accountNumber;
     private String accountHolderName;
     private String accountHolderLastName;
     private String personType;
